refactor(#18): 여행 일정 로드맵 설문 퍼블리싱 컴포넌트 구조 개선
